계좌리스트, 입금, 출금, 이체 페이지

+++ src/main/java/com/tencoding/bank/dto/SaveFormDto.java
@@ -8,6 +8,6 @@ public class SaveFormDto {
 	// 화면 name태그 기준
 	private String number;
 	private String password;
-	private int balance;
+	private Long balance;
 
 }
